requirements.txt
aiohttp
motor
aiohttp_cors
pip
PyJWT